capstone_1/prepare_dataset/requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
imagehash>=4.0
split-folders>=0.5.0